refactor: Complete god component refactoring across all three phases
- Phase 1: Extracted QuizActivity & CustomQuizActivity logic into 4 reusable hooks (useQuizEngine, useQuizState, useQuizSubmission, useQuizUI), reducing components by 34% and 33%
- Phase 2: Decomposed QuizCreation.jsx (592→240 lines, 60% reduction) into 5 focused sub-components for metadata, settings, questions, CSV upload, and difficulty selection
- Phase 3: Refactored MyFlashcards.jsx and QuizList.jsx with filtering hooks and UI components, achieving 58% and 41% reductions respectively with URL sync and debouncing
